fix: 串口接收错误

+++ src/serial/include/serial.hpp
@@ -1,5 +1,6 @@
 #include <serial_driver/serial_driver.hpp>
 #include <iostream>
+
 #include "package.hpp"
 #include "toml.hpp"
 
@@ -11,9 +12,7 @@ using namespace drivers::common;
 class Serial {
 
 public:
-	Serial():
-	io_context(2),
-	serial_driver(io_context) {}
+	Serial(): io_context(2), serial_driver(io_context) {}
 
 	~Serial() {
 		if(serial_driver.port()->is_open()) {
@@ -22,7 +21,7 @@ public:
 	}
 	/**
      * @brief 串口初始化
-     * 
+     *
      * @return 0为正常，1为初始化失败
      */
 	int init();
@@ -30,7 +29,7 @@ public:
 
 	size_t send_target(Data4Send &data);
 
-	size_t receiver(Data4Receive &data);
+	int receiver(Data4Receive &data);
 
 private:
 	std::shared_ptr<SerialPort> port;
